+++ CHANGELOG.md
@@ -123,6 +123,17 @@ in a future release.
   - Normalize parameter descriptive text for default values.
   - BREAKING CHANGE: Database changed to DatabaseName for consistency with
     other modules ([issue #1484](https://github.com/dsccommunity/SqlServerDsc/issues/1484)).
+  - BREAKING CHANGE: The resource no longer create the database user if
+    it does not exist. Use the resource _SqlDatabaseUser_ to enforce that
+    the database user exist in the database prior to setting permissions
+    using this resource ([issue #848](https://github.com/dsccommunity/SqlServerDsc/issues/848)).
+  - BREAKING CHANGE: The resource no longer checks if a login exist so that
+    it is possible to set permissions for database users that does not
+    have a login, e.g. the database user 'guest' ([issue #1134](https://github.com/dsccommunity/SqlServerDsc/issues/1134)).
+  - Updated examples.
+  - Added integration tests ([issue #741](https://github.com/dsccommunity/SqlServerDsc/issues/741)).
+  - Get-TargetResource will no longer throw an exception if the database
+    does not exist.
 - SqlDatabaseRecoveryModel
   - BREAKING CHANGE: The parameter `ServerName` is now non-mandatory and
     defaults to `$env:COMPUTERNAME` ([issue #319](https://github.com/dsccommunity/SqlServerDsc/issues/319)).